fix crafter issues
