+++ src/main/java/me/apeiros/alchimiavitae/setup/items/potions/PotionOfOsmosis.java
@@ -1,107 +1,116 @@
 package me.apeiros.alchimiavitae.setup.items.potions;
 
-import java.util.ArrayList;
 import java.util.Collection;
-import java.util.List;
 
-import org.bukkit.Bukkit;
 import org.bukkit.Color;
+import org.bukkit.Location;
 import org.bukkit.Material;
 import org.bukkit.Particle;
 import org.bukkit.Sound;
+import org.bukkit.World;
 import org.bukkit.entity.Player;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.player.PlayerItemConsumeEvent;
 import org.bukkit.inventory.ItemStack;
-import org.bukkit.inventory.meta.PotionMeta;
 import org.bukkit.potion.PotionEffect;
-import org.jetbrains.annotations.NotNull;
+import org.bukkit.scheduler.BukkitRunnable;
 
 import io.github.thebusybiscuit.slimefun4.api.items.ItemGroup;
-import io.github.thebusybiscuit.slimefun4.core.handlers.ItemUseHandler;
-import io.github.thebusybiscuit.slimefun4.implementation.items.SimpleSlimefunItem;
 
 import me.apeiros.alchimiavitae.AlchimiaUtils;
 import me.apeiros.alchimiavitae.AlchimiaVitae;
 import me.apeiros.alchimiavitae.setup.AlchimiaItems;
+import me.apeiros.alchimiavitae.setup.items.crafters.CosmicCauldron;
 
-public class PotionOfOsmosis extends SimpleSlimefunItem<ItemUseHandler> {
+public class PotionOfOsmosis extends AbstractListenerPotion {
 
-    public PotionOfOsmosis(ItemGroup c) {
-
-        super(c, AlchimiaItems.POTION_OF_OSMOSIS, AlchimiaUtils.RecipeTypes.DIVINE_ALTAR, new ItemStack[] {
+    public PotionOfOsmosis(ItemGroup ig, CosmicCauldron cauldron) {
+        super(ig, AlchimiaItems.POTION_OF_OSMOSIS, AlchimiaUtils.RecipeTypes.DIVINE_ALTAR, new ItemStack[] {
                 AlchimiaItems.EXP_CRYSTAL, new ItemStack(Material.NETHERITE_SCRAP), AlchimiaItems.EXP_CRYSTAL,
                 AlchimiaItems.DARK_ESSENCE, new ItemStack(Material.HONEY_BOTTLE), AlchimiaItems.LIGHT_ESSENCE,
                 AlchimiaItems.DARKSTEEL, new ItemStack(Material.DRAGON_BREATH), AlchimiaItems.ILLUMIUM
-        });
-
+        }, cauldron);
     }
 
-    @NotNull
-    @Override
-    public ItemUseHandler getItemHandler() {
-        return e -> {
-            // Player & item
-            Player p = e.getPlayer();
-            ItemStack potion = e.getItem();
+    // {{{ Handle potion drink
+    @EventHandler(ignoreCancelled = true)
+    public void onDrink(PlayerItemConsumeEvent e) {
+        ItemStack potion = e.getItem();
 
-            // If player has no effects
-            if (p.getActivePotionEffects().isEmpty()) {
-                p.sendMessage(AlchimiaUtils.format("<red>There are no effects to absorb!"));
-                return;
-            }
+        // Make sure the potion is similar to this one
+        if (!this.isSimilar(potion))
+            return;
 
-            // Collection of player's effects
-            Collection<PotionEffect> playerEffectsList = p.getActivePotionEffects();
+        // Cancel the event to prevent the potion from being consumed normally
+        e.setCancelled(true);
 
-            // Remove the player's current potion effects
-            for (PotionEffect eff : playerEffectsList) {
-                p.removePotionEffect(eff.getType());
-            }
+        Player p = e.getPlayer();
+        Collection<PotionEffect> effects = p.getActivePotionEffects();
 
-            // Make a new potion
-            ItemStack newPotion = AlchimiaUtils.makePotion("AV_CORUSCATING_POTION", AlchimiaUtils.format("<gradient:#6fe3e1:#53e6a6>Coruscating Potion</gradient>"), Color.FUCHSIA, playerEffectsList, false);
-            PotionMeta meta = (PotionMeta) newPotion.getItemMeta();
+        // Make sure the player has effects
+        if (p.getActivePotionEffects().isEmpty()) {
+            p.sendMessage(AlchimiaUtils.format("<red>There are no effects to absorb!"));
+            return;
+        }
 
-            // Make a lore
-            List<String> lore = new ArrayList<>();
-            lore.add(AlchimiaUtils.format("<green>Created from a"));
-            lore.add(AlchimiaUtils.format("<gradient:#6274e7:#8752a3>Potion of Osmosis</gradient>"));
-            lore.add("");
-            lore.add(AlchimiaUtils.itemType("Potion"));
+        // Consume Potion of Osmosis
+        p.getInventory().removeItem(potion);
 
-            // Set the lore
-            meta.setLore(lore);
+        // Remove the player's potion effects
+        for (PotionEffect effect : effects) {
+            p.removePotionEffect(effect.getType());
+        }
 
-            // Set item meta
-            newPotion.setItemMeta(meta);
+        // Make a new potion
+        ItemStack newPotion = AlchimiaUtils.makePotion(
+                "AV_CORUSCATING_POTION",
+                AlchimiaUtils.format("<gradient:#6fe3e1:#53e6a6>Coruscating Potion</gradient>"),
+                Color.FUCHSIA, effects, false,
+                "&aCreated from a",
+                AlchimiaUtils.format("<gradient:#6274e7:#8752a3>Potion of Osmosis</gradient>"),
+                "", AlchimiaUtils.itemType("Potion"));
 
-            // Remove potion of osmosis
-            p.getInventory().removeItem(potion);
+        // {{{ Finish
+        new BukkitRunnable() {
+            private int layer = 2;
 
-            // Effects
-            // Layer 0
-            p.getWorld().playSound(p.getLocation(), Sound.BLOCK_BREWING_STAND_BREW, 1, 1);
+            @Override
+            public void run() {
+                World w = p.getWorld();
+                Location l = p.getLocation();
 
-            // Layer 1
-            Bukkit.getScheduler().runTaskLater(AlchimiaVitae.i(), () -> {
-                p.getWorld().playSound(p.getLocation(), Sound.BLOCK_BEACON_ACTIVATE, 1, 1);
+                if (layer == 2) {
+                    // Effects
+                    w.playSound(l, Sound.BLOCK_BREWING_STAND_BREW, 1, 1);
 
-                // Layer 2
-                Bukkit.getScheduler().runTaskLater(AlchimiaVitae.i(), () -> {
-                    p.getWorld().playSound(p.getLocation(), Sound.ITEM_BOTTLE_FILL_DRAGONBREATH, 1, 1);
-                    p.getWorld().playSound(p.getLocation(), Sound.ITEM_BOTTLE_FILL, 1, 1);
-                    p.getWorld().playSound(p.getLocation(), Sound.ITEM_TOTEM_USE, 0.6F, 1);
-                    p.getWorld().spawnParticle(Particle.END_ROD, p.getLocation(), 60, 1, 1, 1);
+                    // Decrease layer
+                    layer--;
+                } else if (layer == 1) {
+                    // Effects
+                    w.playSound(l, Sound.BLOCK_BEACON_ACTIVATE, 1, 1);
 
-                    // Message
-                    p.sendMessage(AlchimiaUtils.format("<green>Successfully bottled your active effects into a potion!"));
+                    // Decrease layer
+                    layer--;
+                } else {
+                    // Effects
+                    w.playSound(l, Sound.ITEM_BOTTLE_FILL_DRAGONBREATH, 1, 1);
+                    w.playSound(l, Sound.ITEM_BOTTLE_FILL, 1, 1);
+                    w.playSound(l, Sound.ITEM_TOTEM_USE, 0.6F, 1);
+                    w.spawnParticle(Particle.END_ROD, l, 60, 1, 1, 1);
 
-                    // Add new potion
+                    // Add potion to player's inventory
                     if (!p.getInventory().addItem(newPotion).isEmpty()) {
                         // Drop it on the ground if no inventory space
-                        p.getWorld().dropItemNaturally(p.getLocation(), newPotion);
+                        w.dropItemNaturally(l, newPotion);
                     }
-                }, 30);
-            }, 30);
-        };
+
+                    // Cancel runnable
+                    this.cancel();
+                }
+            }
+        }.runTaskTimer(AlchimiaVitae.i(), 0, 30);
+        // }}}
     }
+    // }}}
+
 }
